+++ app/emails/consent/invite-catch-up.njk
@@ -35,7 +35,7 @@ Both vaccines are offered to all young people aged 13 to 15 (in school Years 9 a
 
 If you want your child to get this vaccine at school, you need to give consent by {{ session.formatted.firstDate }}.
 
-[Give or refuse consent for the {{ session.vaccination }}](/consents/{{ session.id }})
+[Give or refuse consent for the {{ session.vaccination }}]({{ session.consentUrl }})
 
 Responding will take less than 5 minutes. If you do not give consent, it’s still important for you to let us know.
