+++ app/emails/consent/invite-clinic-consent.njk
@@ -2,7 +2,7 @@ You recently booked {{ consent.child.fullAndPreferredNames }} into a clinic so t
 
 Please give consent for this vaccination ahead of your clinic appointment.
 
-[Give consent for the {{ session.vaccination }}](/consents/{{ session.id }})
+[Give consent for the {{ session.vaccination }}]({{ session.consentUrl }})
 
 ## Contact the local health team
