+++ app/emails/consent/invite-reminder.njk
@@ -20,7 +20,7 @@ The nasal spray contains gelatine. If your child does not use gelatine products,
 
 ## Please give or refuse consent
 
-[Give or refuse consent for the {{ session.vaccination }}](/consents/{{ session.id }})
+[Give or refuse consent for the {{ session.vaccination }}]({{ session.consentUrl }})
 
 Responding will take less than 5 minutes. If you do not give consent, it’s still important for you to respond.
