+++ app/emails/consent/invite-subsequent-reminder.njk
@@ -18,7 +18,7 @@ The nasal spray contains gelatine. If your child does not use gelatine products,
 
 ## Please give or refuse consent
 
-[Give or refuse consent for the {{ session.vaccination }}](/consents/{{ session.id }})
+[Give or refuse consent for the {{ session.vaccination }}]({{ session.consentUrl }})
 
 Responding will take less than 5 minutes.
