+++ app/routes/consent.js
@@ -17,16 +17,4 @@ router.all('/:uuid/add', consentController.readAdd)
 router.get('/:uuid/add', consentController.showAdd)
 router.post('/:uuid/add', consentController.updateAdd)
 
-router.get('/:id/new', consentController.new)
-
-router.all(['/:id', '/:id*'], consentController.read)
-router.get('/:id/:view?', consentController.show)
-
-router.all('/:id/:uuid/?:form(new)/check-answers', consentController.readForm)
-router.post('/:id/:uuid/?:form(new)/check-answers', consentController.update)
-
-router.all('/:id/:uuid/?:form(new|edit)/:view', consentController.readForm)
-router.get('/:id/:uuid/?:form(new|edit)/:view', consentController.showForm)
-router.post('/:id/:uuid/?:form(new|edit)/:view', consentController.updateForm)
-
 export const consentRoutes = router
