+++ app/views/parent/start.njk
@@ -25,7 +25,7 @@
 
       {{ button({
         text: __("consent.start.confirm"),
-        href: "/consents/" + session.id + "/new"
+        href: session.consentUrl + "/new"
       }) }}
 
       <h2 class="nhsuk-heading-m">
